requirements.txt
istari-digital-client>=10.7.0
python-dotenv
